pyproject.toml
[project]
name = "churn-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "harish gehlot", email = "gehloth03@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dagshub>=0.6.3",
    "fastapi[standard]>=0.121.0",
    "hopsworks>=4.4.2",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "just>=0.8.165",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "polars>=1.35.1",
    "python-dotenv>=1.2.1",
    "rust-just>=1.43.0",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
